requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
google-cloud-logging==2.6.0
psycopg2==2.9.1